Use spaces for unallocated character codes
In 8bbacc1d, I switched to using the underscore character as a
placeholder for unallocated PRINTABLE indices, but this has the
disadvantage of causing pprint() to use the underscore for unknown
characters.

I prefer printing a space in those cases (at least for now), so this
change reverts that change to PRINTABLE and adjusts the CHARCODES
comprehension instead.