Rework encode()'s error handling (#22)
We no longer use a regular expression to validate the input string.
Instead, the logic within the per-character encoding loop has been
improved to catch all of the known error cases and raise appropriate
ValueError exceptions.

One notable behavioral change: we now raise an error for unsupported
character codes within the 0..69 range.

This results in slightly faster code with more useful errors. For
example, we now include the actual character that we've identified as
invalid, which aids in debugging.