Add a `max_rows` argument to encode_text() (#115)
This controls the maximum number of rows that will be returned,
defaulting to ROWS (6). It can be set to a lower value to produce a
partial board or 0 to support unlimited rows.

ValueError is no longer raised when the result exceeds max_rows.
Instead, the result is truncated to max_rows.

Also, canonically print the board's dimensions as (COLS, ROWS), which
is a more standard representation.