Add text support to ReadWriteClient.write_message
The Vestaboard Read / Write API now supports passing text in addition to
the previously-supported rows of character codes.

It also now returns HTTP 200 (instead of HTTP 201) upon a successful
write, so update the response logic accordingly.

I've noticed that this endpoint also now returns HTTP 405 when the
written message matches the board's existing message, returning a JSON
error like this:

    {
      'status': 'error',
      'message': 'Message fingerprint matches the message that is already displayed on the board'
    }

This is surfaced to the caller as an HTTPStatusError (just like any
other request failure), but I may look at ways to make this a little
friendlier if this is the official server-side behavior.