Replace httpx with urllib (#237)
We previously used httpx as our HTTP client. It worked well, but our
HTTP needs are very basic, and httpx was our only runtime dependency.

Instead, we now implement a simple HTTP client using urllib from the
Python standard library. The only public-facing change is that we no
longer accept an (optional) http_client instance because that was
httpx-specific.